php comments

+++ src/models/classes/ExpenseList.php
@@ -37,11 +37,8 @@
     //function to save expense:
 		private function saveExpense($args){
       $connection = $args['connection'];
-      $sql="INSERT INTO expense (user_id, amount, category, payment, expense_date, expense_time, location, store, comments)
-      VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)";
-			$stmt = $connection->prepare($sql);
-			$stmt->bind_param("idsssssss", $args['user_id'], $args['amount'], $args['category'], $args['payment'],
-      $args['date'], $args['time'], $args['location'], $args['store'], $args['comments']);
+			$stmt = $connection->prepare("INSERT INTO expense (user_id, amount, category, payment, expense_date, expense_time, location, store, comments)  VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)");
+			$stmt->bind_param("idsssssss", $args['user_id'], $args['amount'], $args['category'], $args['payment'],  $args['date'], $args['time'], $args['location'], $args['store'], $args['comments']);
 			$save_results = array();
 			if($stmt->execute()){
 				$id = $stmt->insert_id;
@@ -54,11 +51,8 @@
 		//function to save expense to 'log_expense' table:
 		private function saveExpenseLog($args){
       $connection = $args['connection'];
-      $sql="INSERT INTO log_expense	(user_id, expense_id, amount, category, payment, expense_date, expense_time, location, store, comments, log_type)
-      VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
-    	$stmt = $connection->prepare($sql);
-			$stmt->bind_param("iidssssssss", $args['user_id'], $args['expense_id'], $args['amount'], $args['category'], $args['payment'],
-      $args['date'], $args['time'], $args['location'], $args['store'], $args['comments'], $args['log_type']);
+    	$stmt = $connection->prepare("INSERT INTO log_expense	(user_id, expense_id, amount, category, payment, expense_date, expense_time, location, store, comments, log_type) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+			$stmt->bind_param("iidssssssss", $args['user_id'], $args['expense_id'], $args['amount'], $args['category'], $args['payment'], $args['date'], $args['time'], $args['location'], $args['store'], $args['comments'], $args['log_type']);
 			$save_results = array();
 			if($stmt->execute()){
 				$id = $stmt->insert_id;
@@ -71,11 +65,8 @@
     //function to update expense:
 		private function updateExpense($args){
       $connection = $args['connection'];
-      $sql="UPDATE expense SET amount=?, category=?, payment=?, expense_date=?, expense_time=?, location=?, store=?, comments=?
-      WHERE id=".$args['expense_id']." AND user_id=".$args['user_id']."";
-      $stmt = $connection->prepare($sql);
-			$stmt->bind_param("dsssssss", $args['amount'], $args['category'], $args['payment'],  $args['date'],
-      $args['time'], $args['location'], $args['store'], $args['comments']);
+      $stmt = $connection->prepare("UPDATE expense SET amount=?, category=?, payment=?, expense_date=?, expense_time=?, location=?, store=?, comments=? WHERE id=".$args['expense_id']." AND user_id=".$args['user_id']."");
+			$stmt->bind_param("dsssssss", $args['amount'], $args['category'], $args['payment'],  $args['date'], $args['time'], $args['location'], $args['store'], $args['comments']);
 			if($stmt->execute()){
 				return true;
 			} else {
@@ -86,9 +77,7 @@
     public function addExpense($args){
       $message = "";
        $connection = $this->db->dbConnect();
-       $categories = $this->db->runQuery("SELECT * FROM `category`
-         WHERE category_name='".$args['category']."' AND (added_by='".$args['user_id']."'
-           OR added_by IS NULL)");
+       $categories = $this->db->runQuery("SELECT * FROM `category` WHERE category_name='".$args['category']."' AND (added_by='".$args['user_id']."' OR added_by IS NULL)");
        $rowcount=mysqli_num_rows($categories);
        if ($rowcount>0) {
          $save_params = $args;
@@ -114,8 +103,7 @@
     public function deleteExpense($args){
       $message = "";
       $connection = $this->db->dbConnect();
-      $deleted = $this->db->deleteFromDB("DELETE FROM `expense` WHERE id='".$args['expense_id']."'
-        AND user_id='".$args['user_id']."'");
+      $deleted = $this->db->deleteFromDB("DELETE FROM `expense` WHERE id='".$args['expense_id']."' AND user_id='".$args['user_id']."'");
       if($deleted===true){
         $save_log_params = $args;
         $save_log_params['connection'] = $connection;
